Web editor: improve delete file
